requirements.txt
# ============================================================
# Harness Engineering - Python 의존성
# pip install -r requirements.txt
# ============================================================

# ─── LLM ────────────────────────────────────────────────────
anthropic>=0.40.0           # Claude API
ollama>=0.4.0               # Ollama (로컬 Gemma)

# ─── GitHub ─────────────────────────────────────────────────
PyGithub>=2.3.0             # GitHub REST API

# ─── Discord 봇 ─────────────────────────────────────────────
discord.py>=2.4.0           # Discord 봇

# ─── HTTP / 비동기 ───────────────────────────────────────────
httpx>=0.27.0
aiohttp>=3.10.0

# ─── 설정 / 유틸 ────────────────────────────────────────────
python-dotenv>=1.0.0        # .env 로드
pydantic>=2.7.0             # 데이터 검증
pydantic-settings>=2.3.0    # 설정 관리
PyYAML>=6.0.1               # YAML 설정 파일

# ─── 스케줄링 ────────────────────────────────────────────────
apscheduler>=3.10.4         # 작업 스케줄러

# ─── 로깅 / 모니터링 ─────────────────────────────────────────
loguru>=0.7.2               # 구조화된 로그
rich>=13.7.0                # 터미널 출력 포매팅

# ─── 개발 도구 ───────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
black>=24.4.0
ruff>=0.4.0
